moved functional API special methods from abstract base container class to the main derived container class.
